pyproject.toml
[project]
name = "otlp-json"
version = "0.9.2"
description = "🐍Lightweight OTEL span to JSON converter, no dependencies, pure Python🐍"
requires-python = ">=3.8"
# https://github.com/astral-sh/uv/issues/4204
# license = "MIT"
readme = "readme.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Software Development :: Libraries",
    "Framework :: OpenTelemetry",
    "Framework :: OpenTelemetry :: Exporters",
]
dependencies = []

[project.urls]
"Repository" = "https://github.com/dimaqq/otlp-json"
"Issues" = "https://github.com/dimaqq/otlp-json/issues"


[dependency-groups]
dev = [
    "opentelemetry-api ~= 1.30",  # 1.30 fixed type hints for Python 3.8
    "typing_extensions ~= 4.12",
    "pyright ~= 1.1",
    "ruff ~= 0.9",
]
testing = [
    "pytest ~= 8.3.4",
    "otlp-test-data >= 0.9.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
